requirements.txt
Office365-REST-Python-Client>=2.5.0
openpyxl>=3.1.0
